requirements.txt
PyQt6==6.6.1
python-docx==1.1.0
openpyxl==3.1.2
jieba==0.42.1
pyinstaller==6.3.0
